[installer]: refactor the mysql incluster into it's own module
