+++ installer/pkg/components/mysql/incluster/constants.go
@@ -2,7 +2,7 @@
 // Licensed under the GNU Affero General Public License (AGPL).
 // See License-AGPL.txt in the project root for license information.
 
-package mysql
+package incluster
 
 import "github.com/gitpod-io/gitpod/installer/pkg/common"
