+++ installer/pkg/components/mysql/incluster/init/00-create-and-init-sessions-db.sql
