+++ installer/pkg/components/mysql/incluster/init/01-recreate-gitpod-db.sql
