+++ installer/pkg/components/mysql/incluster/objects.go
@@ -1,0 +1,16 @@
+// Copyright (c) 2021 Gitpod GmbH. All rights reserved.
+// Licensed under the GNU Affero General Public License (AGPL).
+// See License-AGPL.txt in the project root for license information.
+
+package incluster
+
+import (
+	"github.com/gitpod-io/gitpod/installer/pkg/common"
+)
+
+var Objects = common.CompositeRenderFunc(
+	configmap,
+	secrets,
+	service,
+	common.DefaultServiceAccount(Component),
+)
